backend/requirements-codespaces-rag.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Model inference (CPU-only)
transformers==4.37.0
torch==2.2.2
accelerate==0.25.0
ctransformers==0.2.27  # For GGUF model support
huggingface-hub>=0.19.3

# Vector database and embeddings (Codespaces compatible)
# Using FAISS instead of ChromaDB for better SQLite compatibility
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Document processing
pymupdf==1.23.8  # For PDF text extraction
langchain==0.2.0
langchain-community==0.2.0
langchain-text-splitters==0.2.0

# Utilities
python-dotenv==1.0.0
httpx==0.27.0
urllib3==2.2.0
aiofiles==23.2.1
tenacity==8.2.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1 